fix api timeout typing

+++ src/zenml/integrations/kubernetes/flavors/kubernetes_step_operator_flavor.py
@@ -72,7 +72,7 @@ class KubernetesStepOperatorSettings(BaseSettings):
         "`pod.status.containerStatuses[*].state.waiting.reason` of a job pod, "
         "should cause the job to fail immediately.",
     )
-    api_request_timeout: Optional[float] = Field(
+    api_request_timeout: Optional[int] = Field(
         default=None,
         description="Timeout for API requests in seconds. If not specified, no explicit timeout will be set. ",
     )
